schedule/requirements/base.txt
Django
djangorestframework
django-cors-headers
gunicorn
psycopg2
dj-database-url
dj-static
django-appconf
static3
wsgiref
django_redis
django_cache_url
django-configurations
django-rest-auth
xlsxwriter
celery[redis]
django-extensions
